Added lobby slice in store, added getLobbyInfo api in apis, Added mem…

+++ server/controllers/users.js
@@ -4,8 +4,10 @@ const Lobby = require("../models/lobbyModel");
 exports.createUser = async (req, res) => {
   try {
     const { username, avatar, leader } = req.body;
-    if(!username || !avatar){
-      return res.status(400).json({ message: `All fields are required user : ${username}, avatar: ${avatar}, leader: ${leader}` });
+    if (!username || !avatar) {
+      return res.status(400).json({
+        message: `All fields are required user : ${username}, avatar: ${avatar}, leader: ${leader}`,
+      });
     }
     const user = await User.create({ username, avatar, leader });
     res.status(201).json(user);
@@ -17,6 +19,12 @@ exports.createUser = async (req, res) => {
 exports.leaveLobbyMember = async (req, res) => {
   try {
     const { userId } = req.body;
+    if (!userId) {
+      return res.status(400).json({
+        success: false,
+        message: `User Id is required`,
+      });
+    }
     const user = await User.findById(userId);
     const lobby = await Lobby.findById(user.lobby);
     const updateLobby = await Lobby.findByIdAndUpdate(
@@ -27,23 +35,42 @@ exports.leaveLobbyMember = async (req, res) => {
       .populate("members")
       .exec();
     const deleteMember = await User.findByIdAndDelete(userId);
-    res.status(201).json({ updateLobby });
+    res.status(201).json({
+      success: true,
+      message: "successfully deleted",
+      data: updateLobby,
+    });
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    res.status(500).json({
+      message: error.message,
+      success: false,
+    });
   }
 };
 
 exports.leaveLobbyLeader = async (req, res) => {
   try {
     const { userId } = req.body;
+    if (!userId) {
+      return res.status(400).json({
+        success: false,
+        message: `User Id is required`,
+      });
+    }
     const user = await User.findById(userId);
     const lobby = await Lobby.findById(user.lobby);
     lobby.members.map(async (member) => {
       await User.findByIdAndDelete(member);
     });
     await Lobby.findByIdAndDelete(lobby._id);
-    res.status(201).json({ message: "successfully deleted" });
+    res.status(201).json({
+      success: true,
+      message: "successfully deleted",
+    });
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    res.status(500).json({
+      success: true,
+      message: error.message,
+    });
   }
 };
